# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Loja Controle • Backend P0
  version: 1.0.0
  description: API P0 para integração Admin ↔ Loja.
servers:
  - url: http://localhost:4000
paths:
  /config:
    get:
      summary: Obter configuração da loja
      operationId: getConfig
      tags: [Config]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      summary: Definir/atualizar configuração
      operationId: setConfig
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Config'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /products:
    get:
      summary: Listar produtos
      operationId: listProducts
      tags: [Products]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /products/bulk:
    post:
      summary: Criar/atualizar produtos em lote
      operationId: upsertProductsBulk
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
        '400':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /orders:
    get:
      summary: Listar pedidos
      operationId: listOrders
      tags: [Orders]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      summary: Criar pedido
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /orders/{id}:
    patch:
      summary: Atualizar parcialmente um pedido
      operationId: patchOrder
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/login:
    post:
      summary: Login
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/me:
    get:
      summary: Perfil do usuário autenticado
      operationId: me
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
        '401':
          $ref: '#/components/responses/ErrorResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    Config:
      type: object
      required: [storeName, currency, whatsappNumber]
      properties:
        storeName: { type: string }
        currency: { type: string }
        whatsappNumber: { type: string }
        theme:
          type: object
          additionalProperties: { type: string }
        logoUrl: { type: string, format: uri }
        businessHours:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              open: { type: string }
              close: { type: string }
        deliveryAreas:
          type: array
          items: { type: string }
        paymentMethods:
          type: array
          items: { type: string }
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number, minimum: 0 }
        category: { type: string }
        available: { type: boolean }
        rating: { type: number, minimum: 0, maximum: 5 }
      required: [name, price]
    OrderItem:
      type: object
      properties:
        productId: { type: string }
        qty: { type: integer, minimum: 1 }
        price: { type: number, minimum: 0 }
      required: [productId, qty, price]
    Order:
      type: object
      properties:
        id: { type: string }
        createdAt: { type: string, format: date-time }
        status:
          type: string
          enum: [open, paid, cancelled]
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
      required: [id, createdAt, status, items]
    OrderCreate:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
      required: [items]
    OrderPatch:
      type: object
      properties:
        status:
          type: string
          enum: [open, paid, cancelled]
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
  responses:
    ErrorResponse:
      description: Erro padrão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security: []
tags:
  - name: Config
  - name: Products
  - name: Orders
  - name: Auth